#feat

+++ packages/engine/include/value/integer_variable.hpp
@@ -2,6 +2,6 @@
 #include "simple_variable.hpp"
 namespace atom::engine {
 class integer_variable
-    : public simple_variable<int32_t,
+    : public simple_variable<int64_t,
                              base_variable::variable_type::VT_INTEGER> {};
 } // namespace atom::engine
